Extract AddressTest and NameTest into model.attribute
